backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
filelock==3.12.0
identify==2.5.24
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pre-commit==3.3.2
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
